// src/app/models/interfaces/patient-model.ts
export interface IPatient {
    id: string;
    createdAt: string;
    modifiedAt: string;
    createdById: string;
    createdBy: string;
    modifiedById: string;
    modifiedBy: string;
    accountId: string;
    doctorId: string;
    mrNo: string;
    name: string;
    phoneNumber: string;
    relation: number;
    gender: number;
    age: string;
    registrationDate: string;
    photoPath: string;
  }